Feat: actualizar preliquidado
